[Changed] Naming from namespace to prefix for vulkan related stuff due to implicit namespace conversions which would result in ugly errors [Added] Material Buffer which gets loaded for each model and can be used to index into the the textures

+++ Engine/Include/Core/Render/Resources/Model.hpp
@@ -43,14 +43,14 @@ namespace FS
     
     struct Model
     {
-        std::vector<int> mRootNodes;
+        std::vector<uint64_t> mRootNodes;
         std::vector<Node> mNodes{};
         std::vector<Mesh> mMeshes{};
         std::vector<Material> mMaterials{};
         std::vector<Texture> mTextures{};
 
-        uint64_t mTotalIndicesSize{};
-        uint64_t mTotalVerticesSize{};
+        uint32_t mTotalIndicesSize{};
+        uint32_t mTotalVerticesSize{};
     };
